[MEMO-REFACTOR] 메모 어댑터 ListAdapter 타입으로 변경 및 common 패키지로 이동
MemoAdpater의 경우 다른 Fragment에서 재사용이 가능하기 때문에 common 패키지로 이동함
